[JIT] Add default arguments for class types
**Summary**
This commit adds support for default arguments in methods of class
types. Similar to how default arguments are supported for regular
script functions and methods on scripted modules, default values are
retrieved from the definition of a TorchScript class in Python as Python
objects, converted to IValues, and then attached to the schemas of
already compiled class methods.

**Test Plan**
This commit adds a set of new tests to TestClassType to test default
arguments.

**Fixes**
This commit fixes #42562.

ghstack-source-id: 90b2eeb6560bf223753926e2302f23997c007aa6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45098