Add torch.do_assert, which is symbolically traceable
Summary:

This is a symbolically traceable alternative to Python's `assert`.
It should be useful to allow people who want to use FX to also
be able to assert things.

A bunch of TODO(before) land are inline - would love thoughts
on where is the best place for this code to live, and what this
function should be called (since `assert` is reserved).

Test Plan:

```
python test/test_fx.py TestFX.test_symbolic_trace_assert
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 81c2a41cdb70dd3ea50942231ea5ef64fedd41c6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45188